feat: Add czt_feature_extractor gist and scaled GC bias analysis
This commit introduces the  gist, incorporating several fixes and enhancements:

- **Initial Gist Inclusion**: All files within the  directory are added to version control. This includes source code, data, results, and reports related to DNA breathing dynamics.
- **Data Pipeline Fix**: Addressed a broken data download link for the Brunello library. The  script was updated to correctly parse the newly sourced text file (), generating a comprehensive .
- **Bug Fixes in **:
    - Resolved  in  by properly handling node traversal in the Eulerian path algorithm.
    - Corrected the metric keys used in the final statistics printout to accurately reflect the  output for .
- **Scaled Scenario 1 Validation (GC Bias)**:
    - Implemented  to create a balanced dataset of 1000 sequences (500 high GC, 500 low GC) from the full Brunello library.
    - Executed the GC bias analysis on this scaled dataset.
    - Updated  with the new results, showing that while GC bias remains statistically significant for Peak Magnitude (d=-0.1833, p=0.0020) and Coherence (d=0.1461, p=0.0120), the effect sizes are considerably smaller than initially observed in the n=20 pilot study.
