+++ gists/breathing/czt_feature_extractor/data/processed/brunello_grouped.fasta
@@ -1,0 +1,40 @@
+>2|low_gc|low_gc
+GCTCGTGCTGGAGGTCTATG
+>5|low_gc|low_gc
+GCTGCTGGAGGTCTATGCGT
+>7|low_gc|low_gc
+GCTGGAGGTCTATGCGTGCT
+>8|high_gc|low_gc
+GGAGGTCTATGCGTGCTGGA
+>9|high_gc|low_gc
+GTCTATGCGTGCTGGAGGTC
+>11|low_gc|low_gc
+GCTGCTGGAGGTCTATGCGT
+>14|low_gc|low_gc
+GCTCGTGCTGGAGGTCTATG
+>17|low_gc|low_gc
+GCTGCTGGAGGTCTATGCGT
+>18|high_gc|low_gc
+GGAGGTCTATGCGTGCTGGA
+>19|low_gc|low_gc
+GCTGGAGGTCTATGCGTGCT
+>20|high_gc|high_gc
+GTCTATGCGTGCTGGAGGTC
+>6|high_gc|high_gc
+GCGTGCTGGAGGTCTATGCG
+>10|high_gc|high_gc
+GCGTGCTGGAGGTCTATGCG
+>16|high_gc|high_gc
+GCGTGCTGGAGGTCTATGCG
+>3|high_gc|high_gc
+GCTCTGTGGAGGCGATCGC
+>13|high_gc|high_gc
+GCTCTGTGGAGGCGATCGC
+>4|high_gc|high_gc
+GACGGAGCGAGGGAGGCGAT
+>12|high_gc|high_gc
+GACGGAGCGAGGGAGGCGAT
+>1|high_gc|high_gc
+GGAGCGAGGGAGGCGATCG
+>15|high_gc|high_gc
+GGAGCGAGGGAGGCGATCG
