+++ gists/breathing/czt_feature_extractor/data/processed/brunello_stability.fasta
@@ -1,0 +1,40 @@
+>low_stability_1|1|high_gc
+GGAGCGAGGGAGGCGATCG
+>low_stability_2|15|high_gc
+GGAGCGAGGGAGGCGATCG
+>low_stability_3|4|high_gc
+GACGGAGCGAGGGAGGCGAT
+>low_stability_4|12|high_gc
+GACGGAGCGAGGGAGGCGAT
+>low_stability_5|3|high_gc
+GCTCTGTGGAGGCGATCGC
+>low_stability_6|13|high_gc
+GCTCTGTGGAGGCGATCGC
+>low_stability_7|6|high_gc
+GCGTGCTGGAGGTCTATGCG
+>low_stability_8|10|high_gc
+GCGTGCTGGAGGTCTATGCG
+>low_stability_9|16|high_gc
+GCGTGCTGGAGGTCTATGCG
+>low_stability_10|7|low_gc
+GCTGGAGGTCTATGCGTGCT
+>high_stability_11|19|low_gc
+GCTGGAGGTCTATGCGTGCT
+>high_stability_12|5|low_gc
+GCTGCTGGAGGTCTATGCGT
+>high_stability_13|11|low_gc
+GCTGCTGGAGGTCTATGCGT
+>high_stability_14|17|low_gc
+GCTGCTGGAGGTCTATGCGT
+>high_stability_15|8|high_gc
+GGAGGTCTATGCGTGCTGGA
+>high_stability_16|18|high_gc
+GGAGGTCTATGCGTGCTGGA
+>high_stability_17|9|high_gc
+GTCTATGCGTGCTGGAGGTC
+>high_stability_18|20|high_gc
+GTCTATGCGTGCTGGAGGTC
+>high_stability_19|2|low_gc
+GCTCGTGCTGGAGGTCTATG
+>high_stability_20|14|low_gc
+GCTCGTGCTGGAGGTCTATG
