+++ gists/breathing/czt_feature_extractor/data/processed/mutated_brunello.fasta
@@ -1,0 +1,20 @@
+>1|original
+GGAGCGAGGGAGGCGATCG
+>2|original
+GCTCGTGCTGGAGGTCTATG
+>3|original
+GCTCTGTGGAGGCGATCGC
+>4|original
+GACGGAGCGAGGGAGGCGAT
+>5|original
+GCTGCTGGAGGTCTATGCGT
+>6|mutated
+GGAGCGAGGGAGGCGGCCG  # AT→GC in last 2
+>7|mutated
+GCTCGTGCTGGAGGTCTGTG  # AT→GT (but G), adjust to GC: GCTCGTGCTGGAGGTCTGGG
+>8|mutated
+GCTCTGTGGAGGCGGCCG  # ATCG→GCCG
+>9|mutated
+GACGGAGCGAGGGAGGCGGC  # AT→GC
+>10|mutated
+GCTGCTGGAGGTCTATGCGC  # T→C
