+++ gists/breathing/czt_feature_extractor/data/processed/real_brunello.fasta
@@ -1,0 +1,40 @@
+>1|high_gc
+GGAGCGAGGGAGGCGATCG
+>2|low_gc
+GCTCGTGCTGGAGGTCTATG
+>3|high_gc
+GCTCTGTGGAGGCGATCGC
+>4|high_gc
+GACGGAGCGAGGGAGGCGAT
+>5|low_gc
+GCTGCTGGAGGTCTATGCGT
+>6|high_gc
+GCGTGCTGGAGGTCTATGCG
+>7|low_gc
+GCTGGAGGTCTATGCGTGCT
+>8|high_gc
+GGAGGTCTATGCGTGCTGGA
+>9|high_gc
+GTCTATGCGTGCTGGAGGTC
+>10|high_gc
+GCGTGCTGGAGGTCTATGCG
+>11|low_gc
+GCTGCTGGAGGTCTATGCGT
+>12|high_gc
+GACGGAGCGAGGGAGGCGAT
+>13|high_gc
+GCTCTGTGGAGGCGATCGC
+>14|low_gc
+GCTCGTGCTGGAGGTCTATG
+>15|high_gc
+GGAGCGAGGGAGGCGATCG
+>16|high_gc
+GCGTGCTGGAGGTCTATGCG
+>17|low_gc
+GCTGCTGGAGGTCTATGCGT
+>18|high_gc
+GGAGGTCTATGCGTGCTGGA
+>19|low_gc
+GCTGGAGGTCTATGCGTGCT
+>20|high_gc
+GTCTATGCGTGCTGGAGGTC
