+++ gists/breathing/czt_feature_extractor/data/processed/real_crispr.fasta
@@ -1,0 +1,59 @@
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>high_gc|high_gc
+GCGCGCGCGCGCGCGCGCG
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
+
+>low_gc|low_gc
+ATATATATATATATATATAT
