+++ gists/breathing/czt_feature_extractor/data/processed/sample.fasta
@@ -1,0 +1,6 @@
+>seq1
+ATCGATCGATCGATCGATCG
+>seq2
+GCTAGCTAGCTAGCTAGCTA
+>seq3
+AAATTTCGTTAAATTTCGTT
