+++ gists/breathing/czt_feature_extractor/results/archive/features.csv
@@ -1,0 +1,5 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,20,unknown,10.159896269825122,0,0.08523809523809524,2.7689773303676954,10.159896269825122,0.8989247845118106,10568.562100431367,5.627176686356182,1.0,9.354738115895064,,,,
+1,20,unknown,7.817512558266945,0,0.08523809523809524,2.4698271967676755,7.817512558266945,0.9099258553701894,5959.460099135462,4.192891298920628,1.0,8.64661014591775,,,,
+2,20,unknown,8.435690068060723,0,0.08523809523809524,2.3937637631930735,8.435690068060723,0.9934962264023641,8963.266424928264,5.653267639886895,1.0,8.618203791332725,,,,
+stats,0,stats,,,,,,,,,,,0.0,0.0,0.0,1.0
