+++ gists/breathing/czt_feature_extractor/results/archive/features_test.csv
@@ -1,0 +1,5 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,20,unknown,46.644820993538396,0,0.08523809523809524,2.303183026939784,46.644820993538396,0.8968455269052908,247367.55796217,27.64192170593985,1.0,46.6448209935384,46.6448209935384,,,,
+1,20,unknown,44.209615495217044,0,0.08523809523809524,0.3942110388279773,44.209615495217044,0.9110983722481879,222178.97087728695,26.18217563114899,1.0,44.20961549521705,44.20961549521705,,,,
+2,20,unknown,65.96092612345805,157,0.09755182072829131,1.388904643831069,65.96092612345805,0.9293794659105686,1062246.130459416,64.39480531682497,0.0,54.656057803803606,65.96092612345805,,,,
+stats,0,stats,,,,,,,,,,,,0.0,0.0,0.0,1.0
