+++ gists/breathing/czt_feature_extractor/results/archive/real_features.csv
@@ -1,0 +1,22 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+1,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+2,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001757,0.0,45.56778755420851,45.56778755420851,,,,
+3,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+4,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001757,0.0,45.56778755420851,45.56778755420851,,,,
+5,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+6,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001757,0.0,45.56778755420851,45.56778755420851,,,,
+7,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+8,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001757,0.0,45.56778755420851,45.56778755420851,,,,
+9,19,unknown,45.56778755420851,29,0.0875126050420168,0.9199155656151715,45.56778755420851,0.9483865159342284,396416.16331468005,38.77965327001756,0.0,45.56778755420851,45.56778755420851,,,,
+10,20,unknown,9.243061137595527,0,0.08523809523809524,1.5053632212233001,9.243061137595527,0.8900536444068763,8893.085689109408,5.231850631400546,1.0,9.243061137595534,9.243061137595527,,,,
+11,20,unknown,9.243061137595534,0,0.08523809523809524,1.5053632212232997,9.243061137595534,0.890053644406876,8893.085689109404,5.231850631400546,1.0,9.243061137595529,9.243061137595534,,,,
+12,20,unknown,9.24306113759553,0,0.08523809523809524,1.5053632212233004,9.24306113759553,0.8900536444068763,8893.08568910941,5.2318506314005475,1.0,9.243061137595534,9.243061137595529,,,,
+13,20,unknown,9.243061137595534,0,0.08523809523809524,1.5053632212232997,9.243061137595534,0.890053644406876,8893.085689109404,5.231850631400546,1.0,9.243061137595529,9.243061137595536,,,,
+14,20,unknown,9.24306113759553,0,0.08523809523809524,1.5053632212233004,9.24306113759553,0.8900536444068763,8893.08568910941,5.2318506314005475,1.0,9.243061137595534,9.243061137595529,,,,
+15,20,unknown,9.243061137595534,0,0.08523809523809524,1.5053632212232997,9.243061137595534,0.890053644406876,8893.085689109404,5.231850631400546,1.0,9.243061137595529,9.243061137595536,,,,
+16,20,unknown,9.24306113759553,0,0.08523809523809524,1.5053632212233004,9.24306113759553,0.8900536444068763,8893.08568910941,5.2318506314005475,1.0,9.243061137595534,9.243061137595532,,,,
+17,20,unknown,9.243061137595534,0,0.08523809523809524,1.5053632212232997,9.243061137595534,0.890053644406876,8893.085689109404,5.231850631400546,1.0,9.243061137595529,9.243061137595536,,,,
+18,20,unknown,9.24306113759553,0,0.08523809523809524,1.5053632212233004,9.24306113759553,0.8900536444068763,8893.08568910941,5.2318506314005475,1.0,9.243061137595534,9.243061137595529,,,,
+19,20,unknown,9.243061137595534,0,0.08523809523809524,1.5053632212232997,9.243061137595534,0.890053644406876,8893.085689109404,5.231850631400546,1.0,9.243061137595529,9.243061137595534,,,,
+stats,0,stats,,,,,,,,,,,,0.0,0.0,0.0,1.0
