+++ gists/breathing/czt_feature_extractor/results/real_large_features.csv
@@ -1,0 +1,20 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d_peak_mag,ci_low_peak_mag,ci_high_peak_mag,p_perm_peak_mag,cohens_d_snr,ci_low_snr,ci_high_snr,p_perm_snr,cohens_d_phase_coherence,ci_low_phase_coherence,ci_high_phase_coherence,p_perm_phase_coherence,estimated_power
+0,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,52.23394279315046,59.45642744998262,,,,,,,,,,,,,
+1,20,low_gc,32.71567498603945,255,0.10523809523809523,-0.6126598033451657,32.71567498603945,0.3456138132968646,93031.63290127742,17.760903916147328,1.0,52.52257657284931,32.71567498603946,,,,,,,,,,,,,
+2,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,53.98142772984498,93.22089139296251,,,,,,,,,,,,,
+4,20,low_gc,65.39541871622595,0,0.08523809523809524,0.48459407024850104,65.39541871622595,0.920585070640099,720434.458849573,51.9991177247242,1.0,53.89282315866198,65.39541871622598,,,,,,,,,,,,,
+5,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425039,595362.671966619,44.274815131002036,1.0,53.08349580946174,73.64832398401632,,,,,,,,,,,,,
+6,20,low_gc,41.22263428566286,0,0.08523809523809524,1.790350173245599,41.22263428566286,0.7073910124636531,253515.53205067306,31.146111808089405,1.0,54.99563863736646,41.22263428566287,,,,,,,,,,,,,
+7,20,high_gc,54.786117441684524,0,0.08523809523809524,0.2540929637743384,54.786117441684524,0.9927656162559647,479497.62751797604,42.63350654218464,1.0,48.28736013231657,54.786117441684524,,,,,,,,,,,,,
+8,20,high_gc,58.68965567485921,0,0.08523809523809524,1.2846396620469878,58.68965567485921,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,51.033996033202286,58.68965567485919,,,,,,,,,,,,,
+9,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425037,595362.671966619,44.274815131002036,1.0,55.445597050939206,73.64832398401632,,,,,,,,,,,,,
+10,20,low_gc,65.39541871622593,0,0.08523809523809524,0.4845940702485012,65.39541871622593,0.9205850706400988,720434.458849573,51.99911772472421,1.0,53.488227084478325,65.39541871622595,,,,,,,,,,,,,
+12,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,48.323785685159606,93.22089139296251,,,,,,,,,,,,,
+13,20,low_gc,32.71567498603945,255,0.10523809523809523,-0.6126598033451657,32.71567498603945,0.3456138132968646,93031.63290127742,17.760903916147328,1.0,48.93615926994675,32.71567498603947,,,,,,,,,,,,,
+14,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,58.1203399790878,59.45642744998262,,,,,,,,,,,,,
+15,20,high_gc,73.64832398401633,0,0.08523809523809524,1.2177456881383981,73.64832398401633,0.8729281669425039,595362.6719666192,44.27481513100204,1.0,54.69833093256551,73.64832398401632,,,,,,,,,,,,,
+16,20,low_gc,65.39541871622592,0,0.08523809523809524,0.48459407024850076,65.39541871622592,0.9205850706400988,720434.4588495733,51.999117724724215,1.0,51.70622848985464,65.39541871622592,,,,,,,,,,,,,
+17,20,high_gc,54.786117441684524,0,0.08523809523809524,0.2540929637743383,54.786117441684524,0.9927656162559648,479497.62751797604,42.633506542184634,1.0,45.91964214352292,54.786117441684524,,,,,,,,,,,,,
+18,20,low_gc,41.22263428566286,0,0.08523809523809524,1.7903501732455986,41.22263428566286,0.7073910124636532,253515.53205067303,31.1461118080894,1.0,53.39595759940373,41.22263428566287,,,,,,,,,,,,,
+19,20,high_gc,58.689655674859225,0,0.08523809523809524,1.2846396620469878,58.689655674859225,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,51.670265759241275,58.6896556748592,,,,,,,,,,,,,
+stats,0,stats,,,,,,,,,,,,0.0449705262544176,-1.1806754675477475,1.2943378032968444,0.94,-0.20404722801321878,-1.6925076122800566,0.8163039424403361,0.678,-0.2880439850987602,-1.367638049028625,0.7261804052479738,0.598,0.0
