+++ gists/breathing/czt_feature_extractor/results/scenarios/scenario1_gc_bias/brunello_features.csv
@@ -1,0 +1,20 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,20,low_gc,32.715674986039474,255,0.10523809523809523,-0.6126598033451662,32.715674986039474,0.3456138132968646,93031.63290127725,17.760903916147306,1.0,47.96541321303722,32.715674986039474,,,,
+1,20,low_gc,65.39541871622593,0,0.08523809523809524,0.4845940702485012,65.39541871622593,0.9205850706400988,720434.458849573,51.9991177247242,1.0,50.665631330886576,65.39541871622593,,,,
+2,20,low_gc,41.22263428566286,0,0.08523809523809524,1.7903501732455986,41.22263428566286,0.7073910124636531,253515.532050673,31.1461118080894,1.0,61.17662520257336,41.22263428566286,,,,
+3,20,high_gc,54.78611744168454,0,0.08523809523809524,0.25409296377433843,54.78611744168454,0.9927656162559647,479497.62751797616,42.63350654218464,1.0,46.52663365244909,54.78611744168454,,,,
+4,20,high_gc,58.68965567485921,0,0.08523809523809524,1.2846396620469878,58.68965567485921,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,54.90394312950838,58.6896556748592,,,,
+5,20,low_gc,65.39541871622595,0,0.08523809523809524,0.48459407024850104,65.39541871622595,0.920585070640099,720434.458849573,51.9991177247242,1.0,55.896700137076756,65.39541871622595,,,,
+6,20,low_gc,32.715674986039474,255,0.10523809523809523,-0.6126598033451663,32.715674986039474,0.34561381329686464,93031.63290127725,17.760903916147306,1.0,50.981646209538496,32.715674986039474,,,,
+7,20,low_gc,65.39541871622593,0,0.08523809523809524,0.4845940702485012,65.39541871622593,0.9205850706400988,720434.458849573,51.99911772472421,1.0,52.6677249077587,65.39541871622593,,,,
+8,20,high_gc,54.78611744168454,0,0.08523809523809524,0.25409296377433854,54.78611744168454,0.9927656162559647,479497.627517976,42.633506542184634,1.0,52.07826800148823,54.78611744168454,,,,
+9,20,low_gc,41.22263428566286,0,0.08523809523809524,1.7903501732455986,41.22263428566286,0.7073910124636532,253515.532050673,31.1461118080894,1.0,39.363211019046936,41.22263428566286,,,,
+10,20,high_gc,58.68965567485921,0,0.08523809523809524,1.2846396620469878,58.68965567485921,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,52.13003571942143,58.68965567485922,,,,
+11,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425037,595362.6719666193,44.27481513100205,1.0,52.962146120772275,73.64832398401633,,,,
+12,20,high_gc,73.64832398401633,0,0.08523809523809524,1.2177456881383981,73.64832398401633,0.8729281669425039,595362.6719666192,44.27481513100204,1.0,68.14491370203284,73.64832398401633,,,,
+13,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425039,595362.671966619,44.274815131002036,1.0,67.07456079936034,73.64832398401633,,,,
+14,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,38.379450224871704,93.2208913929625,,,,
+15,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,36.60243146159653,93.22089139296251,,,,
+18,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,62.46903944448309,59.45642744998263,,,,
+19,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,42.84304221137435,59.45642744998263,,,,
+stats,0,stats,,,,,,,,,,,,-1.2936567790628992,-2.7000334660371816,-0.515165267988752,0.0
