+++ gists/breathing/czt_feature_extractor/results/scenarios/scenario1_gc_bias/fixed_test.csv
@@ -1,0 +1,20 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,34.963732827618415,59.45642744998262,,,,
+1,20,low_gc,32.715674986039446,255,0.10523809523809523,-0.6126598033451657,32.715674986039446,0.3456138132968646,93031.63290127742,17.760903916147328,1.0,53.21996140546861,32.715674986039446,,,,
+2,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,50.33602763601313,93.2208913929625,,,,
+4,20,low_gc,65.39541871622595,0,0.08523809523809524,0.48459407024850104,65.39541871622595,0.920585070640099,720434.458849573,51.9991177247242,1.0,45.9733744933349,65.39541871622596,,,,
+5,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425039,595362.671966619,44.274815131002036,1.0,53.37472065839891,73.64832398401634,,,,
+6,20,low_gc,41.22263428566286,0,0.08523809523809524,1.790350173245599,41.22263428566286,0.7073910124636531,253515.53205067306,31.146111808089405,1.0,53.15822069794258,41.22263428566286,,,,
+7,20,high_gc,54.786117441684524,0,0.08523809523809524,0.2540929637743384,54.786117441684524,0.9927656162559648,479497.62751797604,42.63350654218464,1.0,51.839513137404694,54.78611744168452,,,,
+8,20,high_gc,58.68965567485921,0,0.08523809523809524,1.2846396620469878,58.68965567485921,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,49.11412481657641,58.6896556748592,,,,
+9,20,high_gc,73.64832398401633,0,0.08523809523809524,1.217745688138398,73.64832398401633,0.8729281669425039,595362.671966619,44.274815131002036,1.0,62.85574732935485,73.64832398401634,,,,
+10,20,low_gc,65.39541871622593,0,0.08523809523809524,0.4845940702485012,65.39541871622593,0.9205850706400988,720434.458849573,51.9991177247242,1.0,55.42468890964998,65.39541871622593,,,,
+12,19,high_gc,93.2208913929625,0,0.08523809523809524,0.7206652794973747,93.2208913929625,0.9101807590830986,1704166.232163147,80.94823866476662,1.0,52.92775010247582,93.2208913929625,,,,
+13,20,low_gc,32.71567498603945,255,0.10523809523809523,-0.6126598033451657,32.71567498603945,0.3456138132968646,93031.63290127742,17.760903916147328,1.0,44.69668913216889,32.71567498603945,,,,
+14,19,high_gc,59.45642744998263,0,0.08523809523809524,1.3786490094530655,59.45642744998263,0.7827436919531185,687306.0591225177,51.6846731246773,1.0,46.6654661182806,59.45642744998262,,,,
+15,20,high_gc,73.64832398401633,0,0.08523809523809524,1.2177456881383981,73.64832398401633,0.8729281669425039,595362.6719666192,44.27481513100204,1.0,54.13510639937486,73.64832398401634,,,,
+16,20,low_gc,65.39541871622592,0,0.08523809523809524,0.48459407024850076,65.39541871622592,0.9205850706400988,720434.4588495733,51.999117724724215,1.0,57.828835641877035,65.39541871622592,,,,
+17,20,high_gc,54.786117441684524,0,0.08523809523809524,0.2540929637743383,54.786117441684524,0.9927656162559648,479497.62751797604,42.633506542184634,1.0,40.26358693969023,54.78611744168452,,,,
+18,20,low_gc,41.22263428566286,0,0.08523809523809524,1.7903501732455986,41.22263428566286,0.7073910124636532,253515.53205067303,31.1461118080894,1.0,60.42720848121271,41.22263428566286,,,,
+19,20,high_gc,58.689655674859225,0,0.08523809523809524,1.2846396620469878,58.689655674859225,0.9848844359090271,542284.2143605473,45.262930359532135,1.0,48.594741626205106,58.68965567485922,,,,
+stats,0,stats,,,,,,,,,,,,0.0,0.0,0.0,1.0
