--- gists/breathing/czt_feature_extractor/sample.fasta
@@ -1,6 +1,0 @@
->seq1
-ATCGATCGATCGATCGATCG
->seq2
-GCTAGCTAGCTAGCTAGCTA
->seq3
-AAATTTCGTTAAATTTCGTT
