+++ gists/breathing/czt_feature_extractor/src/features.csv
@@ -1,0 +1,5 @@
+seq_id,seq_len,label,peak_mag,peak_idx,peak_freq,peak_phase,peak_to_skirt,phase_coherence,band_energy,snr,qc_flag,shuffle_mean_peak,phase_scramble_mean_peak,cohens_d,ci_low,ci_high,p_perm
+0,20,unknown,46.64482099353834,0,0.08523809523809524,2.303183026939784,46.64482099353834,0.896845526905292,247367.5579621698,27.641921705939843,1.0,46.64482099353836,46.64482099353834,,,,
+1,20,unknown,44.209615495217015,0,0.08523809523809524,0.3942110388279769,44.209615495217015,0.9110983722481882,222178.970877287,26.182175631148993,1.0,44.20961549521704,44.20961549521702,,,,
+2,20,unknown,65.96092612345804,157,0.09755182072829131,1.3889046438310688,65.96092612345804,0.9293794659105686,1062246.130459416,64.39480531682497,0.0,54.65605780380359,65.96092612345805,,,,
+stats,0,stats,,,,,,,,,,,,0.0,0.0,0.0,1.0
